Inventory upgrade
Further improved inventory and also realised all actions will have to be moved into rooms.py as they will be used there. Also every time there will be a change of "scene", ..._entry() the player will have to be added in as parameter so everything will have access to it.